while true works, batch limit order creation not yet checked
