+++ utils/__init__.py
@@ -1,3 +1,4 @@
 from .utils import *
 from .config_setup import *
 from .error_checkers import *
+from .sides_directions import *
