+++ utils/utils.py
@@ -1,7 +1,6 @@
 import pytz
 
 from datetime import datetime, timedelta
-from utils import GridDirections
 
 
 # example: datetime.strptime("2023-07-01 00:00:00", "%Y-%m-%d %H:%M:%S")
@@ -24,24 +23,6 @@ def add_get_attributes(cls):
 
     cls.get_attributes = get_attributes
     return cls
-
-
-def get_side_from_direction(direction):
-    if direction == GridDirections.GRID_DIRECTION_LONG:
-        return 'long'
-    elif direction == GridDirections.GRID_DIRECTION_SHORT:
-        return 'short'
-    else:
-        raise ValueError('Invalid grid direction')
-
-
-def get_direction_from_side(side):
-    if side == 'long':
-        return GridDirections.GRID_DIRECTION_LONG
-    elif side == 'short':
-        return GridDirections.GRID_DIRECTION_SHORT
-    else:
-        raise ValueError('Invalid grid direction')
 
 
 if __name__ == '__main__':
